requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
beautifulsoup4==4.13.5
    # via -r requirements.in
build==1.3.0
    # via pip-tools
camelot-py==1.0.9
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
chardet==5.2.0
    # via camelot-py
charset-normalizer==3.4.3
    # via
    #   pdfminer-six
    #   requests
click==8.3.0
    # via
    #   camelot-py
    #   pip-tools
    #   uvicorn
coverage[toml]==7.10.7
    # via pytest-cov
cryptography==46.0.1
    # via pdfminer-six
distro==1.9.0
    # via
    #   openai
    #   tabula-py
et-xmlfile==2.0.0
    # via openpyxl
execnet==2.1.1
    # via pytest-xdist
faiss-cpu==1.12.0
    # via -r requirements.in
fastapi==0.117.1
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via -r requirements.in
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via -r requirements.in
jiter==0.11.0
    # via openai
joblib==1.5.2
    # via scikit-learn
lxml==6.0.2
    # via python-docx
markupsafe==3.0.2
    # via jinja2
mypy==1.18.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
numpy==2.2.6
    # via
    #   -r requirements.in
    #   camelot-py
    #   faiss-cpu
    #   opencv-python-headless
    #   pandas
    #   scikit-learn
    #   scipy
    #   tabula-py
openai==1.109.0
    # via -r requirements.in
opencv-python-headless==4.12.0.88
    # via camelot-py
openpyxl==3.1.5
    # via
    #   -r requirements.in
    #   camelot-py
opentelemetry-api==1.37.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.37.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.37.0
    # via -r requirements.in
opentelemetry-proto==1.37.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.37.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   build
    #   faiss-cpu
    #   pytesseract
    #   pytest
pandas==2.3.2
    # via
    #   -r requirements.in
    #   camelot-py
    #   tabula-py
pathspec==0.12.1
    # via mypy
pdfminer-six==20250506
    # via
    #   -r requirements.in
    #   camelot-py
pillow==11.3.0
    # via
    #   -r requirements.in
    #   camelot-py
    #   pytesseract
pip-tools==7.5.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
protobuf==6.32.1
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pymupdf==1.26.4
    # via -r requirements.in
pypdf==5.9.0
    # via camelot-py
pypdfium2==4.30.0
    # via camelot-py
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytesseract==0.3.13
    # via -r requirements.in
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-sugar
    #   pytest-xdist
pytest-cov==7.0.0
    # via -r requirements.in
pytest-sugar==1.1.1
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via -r requirements.in
rapidfuzz==3.14.1
    # via -r requirements.in
regex==2025.9.18
    # via tiktoken
requests==2.32.5
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   tiktoken
ruff==0.13.1
    # via -r requirements.in
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.16.2
    # via scikit-learn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
starlette==0.48.0
    # via fastapi
structlog==25.4.0
    # via -r requirements.in
tabula-py==2.10.0
    # via -r requirements.in
tabulate==0.9.0
    # via camelot-py
tenacity==9.1.2
    # via -r requirements.in
termcolor==3.1.0
    # via pytest-sugar
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.11.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
types-pyyaml==6.0.12.20250915
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   mypy
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.37.0
    # via -r requirements.in
uvloop==0.21.0 ; sys_platform != "win32"
    # via -r requirements.in
watchfiles==1.1.0
    # via -r requirements.in
websockets==15.0.1
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
